+++ Source/Patches/QuotaPatch.cs
@@ -5,7 +5,7 @@ namespace ConfigurableWarning.Patches {
         [HarmonyPostfix]
         [HarmonyPatch(typeof(RoomStatsHolder), MethodType.Constructor, [typeof(SurfaceNetworkHandler), typeof(int), typeof(int), typeof(int)])]
         internal static void Constructor(RoomStatsHolder __instance, SurfaceNetworkHandler handler, int startMoney, int startQuotaToReachToReach, int daysPerQuota) {
-            AccessTools.Property(typeof(RoomStatsHolder), "DaysPerQutoa").SetValue(__instance, Plugin.Instance.config.daysPerQuota.Value);
+            AccessTools.Property(typeof(RoomStatsHolder), "DaysPerQutoa").SetValue(__instance, Plugin.Instance.PluginConfig.daysPerQuota.Value);
         }
     }
 }