+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2024-04-12
+
+### Changed
+
+- Reorganized settings to make it easier to work with
+- Fixed a few issues
+- Made the patches better
+- We no longer rely on ContentSettings!
+
+### Added
+
+- Modified ContentSettings API is now built-in
+- Settings patch
+- `TextSetting` and `TextSettingsUI` for blank "settings" (basically a label/divider)
+- `Common*Setting` classes for ease
+- Basic register & setup function for settings (DX stuff)
+
 ## [1.6.0] - 2024-04-11
 
 ### Added